- All Media Collection Views now use UICollectionView instead of LazyStacks for performance - DebugMode now works as expected - SortViews (all groupings are now handled in the UICollectionView itself)
