+++ ABJC/UI/LibraryView/MediaCollectionView.swift
@@ -26,37 +26,7 @@ extension LibraryView
             self.type = type
         }
         
-        let new = true
-        
-        var body: some View {
-            if new {
-                newBody
-            } else {
-                oldBody
-            }
-        }
-        
-        var oldBody: some View
-        {
-            NavigationView {
-                ScrollView(.vertical, showsIndicators: true) {
-                    if items.count != 0, let items = items {
-                        GroupingViewContainer(items).environmentObject(session)
-                    } else {
-                        ActivityIndicatorView()
-                    }
-                }.edgesIgnoringSafeArea(.horizontal)
-            }
-            .onAppear(perform: load)
-            // Present MediaPlayer when itemPlaying is pending
-            .fullScreenCover(item: $session.itemPlaying, onDismiss: session.restoreFocus) { item in
-                MediaPlayerView(item)
-                    .environmentObject(session)
-            }
-        }
-        
-        
-        var newBody: some View
+        var body: some View
         {
             NavigationView {
                 if items.count != 0, let items = items {
