+++ ABJC/en.lproj/Localizable.stringsdict
@@ -1,0 +1,172 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>postertype.wide</key>
+	<string>Breit (Backdrop)</string>
+	<key>postertype.poster</key>
+	<string>Poster</string>
+	<key>reportaproblem.title</key>
+	<string>Report A Problem</string>
+	<key>pref.debugmenu.debugmode.label</key>
+	<string>Debug Modus</string>
+	<key>pref.debugmenu.alerts.auth</key>
+	<string>Authentifizierung</string>
+	<key>pref.debugmenu.alerts.api</key>
+	<string>API</string>
+	<key>pref.debugmenu.alerts.playback</key>
+	<string>Playback</string>
+	<key>pref.debugmenu.clearimagecache.label</key>
+	<string>Bild-Cache löschen</string>
+	<key>pref.client.tabs.disclaimer</key>
+	<string>UI wird nach einem Neustart geupdated</string>
+	<key>pref.client.collectiongrouping.label</key>
+	<string>Collection Grouping</string>
+	<key>collectionGrouping.genre</key>
+	<string>Genre</string>
+	<key>collectionGrouping.releaseyear</key>
+	<string>Year</string>
+	<key>collectionGrouping.releasedecade</key>
+	<string>Decade</string>
+	<key>collectionGrouping.title</key>
+	<string>Title</string>
+	<key>library.search.results</key>
+	<string>Results</string>
+	<key>pref.client.tabs.watchnow.descr</key>
+	<string>favorites &amp; continue now</string>
+	<key>pref.client.tabs.series.descr</key>
+	<string>all series you have access to</string>
+	<key>pref.client.tabs.movies.descr</key>
+	<string>all movies you have access to</string>
+	<key>pref.client.tabs.search.descr</key>
+	<string>alphabetical list of library &amp; search</string>
+	<key>pref.client.showstitles.descr</key>
+	<string>Show a titles for items</string>
+	<key>pref.client.postertype.descr</key>
+	<string>You can choose between wide 16:9 posters or the classic 2:3 poster</string>
+	<key>pref.client.postertype.label</key>
+	<string>Type of Poster for Media Items</string>
+	<key>pref.client.showstitles.label</key>
+	<string>Always Show Titles</string>
+	<key>pref.client.ui.label</key>
+	<string>User Interface</string>
+	<key>pref.general.label</key>
+	<string>General</string>
+	<key>pref.userinfo.id</key>
+	<string>id</string>
+	<key>pref.userinfo.name</key>
+	<string>username</string>
+	<key>pref.userinfo.label</key>
+	<string>User</string>
+	<key>pref.libraryinfo.label</key>
+	<string>Library</string>
+	<key>pref.libraryinfo.episodes</key>
+	<string>episodes</string>
+	<key>pref.libraryinfo.series</key>
+	<string>series</string>
+	<key>pref.libraryinfo.total</key>
+	<string>total</string>
+	<key>pref.libraryinfo.movies</key>
+	<string>movies</string>
+	<key>betaflags.singlepagemode.descr</key>
+	<string>Show Library as single page rather than multiple tabs</string>
+	<key>betaflags.singlepagemode.label</key>
+	<string>Single Page Mode</string>
+	<key>betaflags.uglymode.descr</key>
+	<string>Fast but Ugly: Fugly</string>
+	<key>betaflags.uglymode.label</key>
+	<string>Fugly Mode</string>
+	<key>pref.client.label</key>
+	<string>Client</string>
+	<key>pref.debugmenu.label</key>
+	<string>Debug Menu</string>
+	<key>pref.client.betaflags.label</key>
+	<string>Beta Options</string>
+	<key>pref.client.tabs.search.label</key>
+	<string>Search</string>
+	<key>pref.client.tabs.series.label</key>
+	<string>TV Shows</string>
+	<key>pref.client.tabs.movies.label</key>
+	<string>Movies</string>
+	<key>pref.client.tabs.watchnow.label</key>
+	<string>Watch Now</string>
+	<key>pref.client.tabs.label</key>
+	<string>Tabs</string>
+	<key>itemdetail.people.label</key>
+	<string>People</string>
+	<key>itemdetail.recommended.label</key>
+	<string>Similar</string>
+	<key>pref.serverinfo.networking</key>
+	<string>networking</string>
+	<key>pref.serverinfo.address</key>
+	<string>address</string>
+	<key>pref.serverinfo.architecture</key>
+	<string>system architecture</string>
+	<key>pref.serverinfo.os</key>
+	<string>operating system</string>
+	<key>pref.serverinfo.id</key>
+	<string>id</string>
+	<key>pref.serverinfo.version</key>
+	<string>version</string>
+	<key>pref.serverinfo.servername</key>
+	<string>name</string>
+	<key>pref.serverinfo.general</key>
+	<string>General</string>
+	<key>buttons.logout</key>
+	<string>Logout</string>
+	<key>pref.serverinfo.label</key>
+	<string>Server</string>
+	<key>library.preferences.label</key>
+	<string>Preferences</string>
+	<key>library.search.label</key>
+	<string>Search</string>
+	<key>library.search.descr</key>
+	<string>Search your library</string>
+	<key>library.series.label</key>
+	<string>TV Shows</string>
+	<key>library.series.descr</key>
+	<string>Display all your TV Shows</string>
+	<key>library.movies.label</key>
+	<string>Movies</string>
+	<key>library.movies.descr</key>
+	<string>Display all your Movies</string>
+	<key>library.watchnow.label</key>
+	<string>Watch Now</string>
+	<key>library.watchnow.descr</key>
+	<string>Continue where you left off and check out your favorites / latest items</string>
+	<key>alerts.auth.failed.label</key>
+	<string>Please check if the username and password is correct</string>
+	<key>alerts.auth.title</key>
+	<string>Authentication Error</string>
+	<key>alerts.info.title</key>
+	<string>Information</string>
+	<key>authView.credentialEntryView.title</key>
+	<string>Credentials</string>
+	<key>buttons.signin</key>
+	<string>sign in</string>
+	<key>authView.credentialEntryView.isUsingHTTPS.label</key>
+	<string>HTTPS</string>
+	<key>authView.credentialEntryView.password.label</key>
+	<string>password</string>
+	<key>authView.credentialEntryView.username.label</key>
+	<string>username</string>
+	<key>buttons.continue</key>
+	<string>continue</string>
+	<key>alerts.removeServer.label</key>
+	<string>Are you sure you want to remove the server?</string>
+	<key>alerts.removeServer.title</key>
+	<string>Remove Server</string>
+	<key>authView.serverSelection.port.path</key>
+	<string>/jellyfin</string>
+	<key>authView.serverSelection.port.label</key>
+	<string>port</string>
+	<key>authView.serverSelection.host.label</key>
+	<string>host</string>
+	<key>authView.serverSelection.manual.descr</key>
+	<string>enter address manually</string>
+	<key>authView.serverSelection.manual.label</key>
+	<string>Manual Selection</string>
+	<key>authView.serverSelection.title</key>
+	<string>Server Selection</string>
+</dict>
+</plist>
