+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.0.0-build.26] - [2021-07-14]
+### Added
+### Changed
+- All Media Collection Views now use UICollectionView instead of LazyStacks for performance
+### Deprecated
+### Fixed 
+- DebugMode now works as expected
+### Removed
+- SortViews (all groupings are now handled in the UICollectionView itself)
+
 ## [v1.0.0-build.25] - [2021-06-27]
 ### Added
 - Added User Selection View
